+++ src/config/constant.dist.php
@@ -1,0 +1,3 @@
+<?php
+define('PDFIMAGE_ENV', 'prod');
+define('PDFIMAGE_ERROR_REPORTING', 0);